requirements.txt
alembic==1.0.0
asn1crypto==0.24.0
Babel==2.6.0
bcrypt==3.1.4
blinker==1.4
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3.1
coverage==4.5.1
dj-database-url==0.5.0
Django==2.1.1
django-cors-headers==2.4.0
django-filter==2.0.0
django-heroku==0.3.1
djangorestframework==3.8.2
extras==1.0.0
fixtures==3.0.0
Flask==0.12.2
Flask-BabelEx==0.9.3
Flask-Gravatar==0.5.0
Flask-HTMLmin==1.3.2
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.9.0
html5lib==1.0.1
htmlmin==0.1.12
httplib2==0.9.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
keyring==10.6.0
keyrings.alt==3.0
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
linecache2==1.0.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
oauth==1.0.1
olefile==0.45.1
paramiko==2.4.2
passlib==1.7.1
pbr==3.1.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pyasn1==0.4.4
pycairo==1.16.2
pycparser==2.18
pycrypto==2.6.1
pylint==2.1.1
pymacaroons==0.13.0
PyNaCl==1.2.1
pyRFC3339==1.0
pyrsistent==0.14.2
pytest==3.8.0
pytest-cov==2.6.0
pytest-pylint==0.12.3
python-dateutil==2.7.3
python-debian==0.1.32
python-editor==1.0.3
python-mimeparse==1.6.0
python-memcached==1.59
pytz==2018.5
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
speaklater==1.3
SQLAlchemy==1.2.11
sqlparse==0.2.4
sshtunnel==0.1.4
stevedore==1.29.0
testtools==2.3.0
traceback2==1.4.0
unittest2==1.1.0
urllib3==1.22
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
wadllib==1.3.2
webencodings==0.5.1
Werkzeug==0.14.1
whitenoise==4.1
WTForms==2.1
zope.interface==4.3.2
iron-cache==0.3.2
iron-core==1.2.0
